Adjustments
preprocessing changed to make block if signal happens 28+16 before hit. The reson is that our "zero" hit for SP has to be placed at 28 ticks after last measurement and we do not it to intercect with treshold crossing for next block

solver2D changed. We fix start time at threshold crossing for now. We also do not subtract actual measurement at the time of measurement in toy current. instead we move it to zero including one more time tick for dead time.

Update README.org